Correctly handle Periods made of Dates
If the Period is made of Dates, then we need to subtract a full day off the end of each split Period since that is the smallest unit of time on a Date.

This is to ensure that any Date in the original Period appears in exactly one of the returned Periods.