Add setuptools as build dependency on Python 3.12+
Python 3.12 implements PEP 632 to remove distutils from the standard
library, but Pendulum 2.x's build script depends on the module to build
the C extension. This implements the lowest effort solution, adding
setuptools as a build dependency on Python 3.12 and later, which
injects distutils automatically back into the environment for
compatibility.